Move one step into KlarnaExpressCheckout folder
